Code cleanup now that the <clinit> bug has been dodged.
